+++ src/Framework/App.php
@@ -8,17 +8,17 @@ class App
 
     const HOOK_HTTP_404                    = "parable_http_404";
     const HOOK_HTTP_200                    = "parable_http_200";
-    const HOOK_RESPONSE_SEND               = "parable_response_send";
+    const HOOK_INIT_DATABASE_BEFORE        = "parable_init_database_before";
+    const HOOK_INIT_DATABASE_AFTER         = "parable_init_database_after";
     const HOOK_LOAD_INITS_AFTER            = "parable_load_inits_after";
+    const HOOK_LOAD_ROUTES_BEFORE          = "parable_load_routes_before";
+    const HOOK_LOAD_ROUTES_NO_ROUTES_FOUND = "parable_load_routes_no_routes_found";
+    const HOOK_LOAD_ROUTES_AFTER           = "parable_load_routes_after";
+    const HOOK_RESPONSE_SEND               = "parable_response_send";
     const HOOK_ROUTE_MATCH_BEFORE          = "parable_route_match_before";
     const HOOK_ROUTE_MATCH_AFTER           = "parable_route_match_after";
     const HOOK_SESSION_START_BEFORE        = "parable_session_start_before";
     const HOOK_SESSION_START_AFTER         = "parable_session_start_after";
-    const HOOK_LOAD_ROUTES_BEFORE          = "parable_load_routes_before";
-    const HOOK_LOAD_ROUTES_NO_ROUTES_FOUND = "parable_load_routes_no_routes_found";
-    const HOOK_LOAD_ROUTES_AFTER           = "parable_load_routes_after";
-    const HOOK_INIT_DATABASE_BEFORE        = "parable_init_database_before";
-    const HOOK_INIT_DATABASE_AFTER         = "parable_init_database_after";
 
     /** @var \Parable\Framework\Config */
     protected $config;